requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
distlib==0.3.6
Django==3.2.2
django-phonenumber-field==5.1.0
django-redis==4.12.1
django-rest-passwordreset==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.8.0
huey==2.3.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
phonenumbers==8.12.21
Pillow==8.2.0
platformdirs==2.5.4
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dotenv==0.17.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
twilio==6.57.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.4
virtualenv==20.16.7
zope.interface==5.4.0
